[FEATURE] Display space names in contact list
